[SEN-10] feat : 패키지 목록 반환

+++ sena-back/src/main/java/org/example/admin/dto/ItemResponseDto.java
@@ -11,6 +11,6 @@ public class ItemResponseDto {
 
     private Long itemId;
     private String itemName;
-    private Long ruby;
+    private Double ruby;
     private String img;
 }
